feat: convert to tsx
